Extract AudioInputManager from SoundTagManager
